Separate code relating to referenced fields out of the main file
Change-type: patch
